Make the async APIs thread-safe and fix the indexes they report (#41)
* Make the asynchronous APIs thread-safe

Concurrent writes to arrays are not safe, resulting in this error
https://stackoverflow.com/questions/40080508/swift-unsafemutablepointer-deinitialize-fatal-error-with-negative-count-when-ap
and others.

* Fix the indexes reported by the asynchronous APIs

The asynchronous APIs parallelize their search over chunks of the list being
searched. These parallel searches previously reported the indexes of the
matching results as relative to the chunk being searched. This change
adjusts the index by the offset of the chunk within the overall list.